pyproject.toml
[project]
name = "DL-boilerplate"
version = "0.1.0"
description = "Modern Deep Learning boilerplate with PyTorch Lightning, Hydra, and FastAPI"
authors = [{name = "Anwesh", email = "anwesh.marwade@pon.com"}]
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    # Core ML and DL libraries
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pytorch-lightning>=2.0.0",
    "torchmetrics>=1.0.0",
    # Configuration management
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    # API framework
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    # Experiment tracking and MLOps
    "clearml>=1.13.0",
    # Model providers and utilities
    "transformers>=4.30.0",
    "ultralytics>=8.0.0",
    # Data handling and utilities
    "pandas>=2.0.3",
    "numpy>=1.24.0",
    "pillow>=9.5.0",
    "opencv-python>=4.8.0",
    # Visualization and monitoring
    "matplotlib>=3.7.2",
    "seaborn>=0.12.2",
    "tensorboard>=2.13.0",
    "wandb>=0.15.0",
    # Development and utilities
    "loguru>=0.7.0",
    "tqdm>=4.65.0",
    "pydantic>=2.0.3",
    "python-multipart>=0.0.6",
    # Data science and ML utilities
    "scikit-learn>=1.3.0",
    "albumentations>=1.3.1",
    # Development tools
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "pytest>=7.4.0",
    "pre-commit>=3.3.0",
    # Legacy dependencies
    "jupyter>=1.0.0",
    "grad-cam>=1.4.8",
    "rich>=14.1.0",
    "ruamel-yaml>=0.18.14",
]



[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
